cros-workon: disable automatic `unpack` when using git-2_src_unpack
If cros-workon can find the git tree it needs locally, it'll copy sources
to the right place by itself.  When it can't, it falls back to using the
git-2.eclass to do the fetch+unpack.  This latter func will automatically
run `unpack` for us which doesn't happen otherwise.

Combine this with minilayout, we have a bunch of packages which implicitly
have `unpack` run for them.  Usually SRC_URI isn't also set which means
nothing happens.  But the xf86-video-armsoc incidentally has SRC_URI set
to an old version of source.  This gets unpacked over top of the git tree
we just unpacked.

Since we don't want this behavior, and few people ever see it (since this
only comes up in a minilayout setup), disable the automatic unpack.

BUG=chromium-os:38342
TEST=`cbuildbot {x86,amd64,arm}-generic-full` works
TEST=`emerge-daisy xf86-video-armsoc` works w/minilayout

Change-Id: Ic72ea76580b46f41bb6a5603d98ffbc6940682ea
Reviewed-on: https://gerrit.chromium.org/gerrit/42703
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>