dev-install: fix emerge setup for private boards
Installing dev-install for private overlays often results in (non-fatal
errors like this:
...
[blocks B     ] chromeos-base/chromeos-firmware-null ("chromeos-base/chromeos-firmware-null"
                    is blocking chromeos-base/chromeos-firmware-daisy-0.0.1-r49)

 * Error: The above package list contains packages which cannot be
 * installed at the same time on the same system.

  (chromeos-base/chromeos-firmware-null-0.0.2-r1::chromiumos, ebuild scheduled for merge to '/build/daisy/') pulled in by
    chromeos-base/chromeos-firmware-null required by (virtual/chromeos-firmware-1::chromiumos, ebuild scheduled for merge to '/build/daisy/')

  (chromeos-base/chromeos-firmware-daisy-0.0.1-r49::daisy-private, installed in '/build/daisy/') pulled in by
    chromeos-base/chromeos-firmware-daisy required by @selected
...

This is because portage modifies certain env vars before running the
ebuild such as PORTDIR_OVERLAY.  In this case, it sets it to only
include the current overlay rather than all of them.  That prevents
the sub-emerge from searching the private board overlays and finding
the packages it requires.

Change the code to clear the env before running these commands.  The
sub-emerge will bootstrap itself to use all the right values.  Also
fix it to use the new multiprocessing eclass and throw an error when
things fail.

BUG=None
TEST=`emerge-daisy dev-install` no longer dumps errors, and installed files are much more complete

Change-Id: Id96ff53b0fe672451a19b1977fdcb23e98cf4cf1
Reviewed-on: https://gerrit.chromium.org/gerrit/42991
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>