dev-install: pass down USE=cros-debug setting
The build_packages script will often change the cros-debug USE flag via
the env, so when we run our sub-emerges, it can easily get a conflicting
setting which results in dependency errors.  Pass down this flag to avoid
that.

Note that this can happen with any USE flag.  We single out this guy as
the rest of our build scripts are already doing this.  With newer portage
behavior (filtering the USE env on a per-package basis based on IUSE), I
can't think of a better solution atm.

BUG=None
TEST=`emerge-x86-alex dev-install` no longer fails when USE=cros-debug or USE=-cros-debug

Change-Id: I6592b56cb67f2ef3444a5f173dd574897adc99df
Reviewed-on: https://gerrit.chromium.org/gerrit/43194
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>