Let the vboot_reference Makefile install the utilities.
The ebuild has been pulling various bits of the vboot utilities out of the
build directory and shoving them random places. That's really the job of the
Makefile's "install" target. This cleans it up a bit.

Note that there's a simultaneous CL for the vboot Makefile that must go in
at the same time.

BUG=chromium-os:37062
BRANCH=none
TEST=none
CQ-DEPEND=CL:43237

This is just a refactoring of the Makefile/ebuild interaction. The end
result should be the same: install the vboot utilities in the correct places
on both host and target.

I ran:

  sudo emerge vboot_reference
  emerge-link vboot_reference

And inspected every file that was touched as a result. I ran a trybot on
daisy-paladin link-paladin lumpy-paladin, too. Seems to be working.

Change-Id: Ie61fb078d2658bb2c99cc935630da9ff623391c4
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43236
Reviewed-by: Randall Spangler <rspangler@chromium.org>