Fix cros-devutils ebuild to run tests properly
The tests fail because devserver_test.py file does not exist anymore
(the file was renamed without updating the ebuild)

With this change the tests pass again.

Also, note that many unittests invoke doctest.testmod() but do not
return the return value as the exit code. Which means that even if
these tests fail, the FEATURE=test ebuild invocation still reports
success. This will have to be handled under a different CL.

BUG=None
TEST=see below
	FEATURES=test sudo -E emerge cros-devutils

Change-Id: I9b9f91a75008daea5f00e65585c7c4b31e0e37af
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43255
Reviewed-by: Mike Frysinger <vapier@chromium.org>