cros-workon: disallow S=$WORKDIR builds
If you try to build dev-install or xorg-conf with a minilayout, it'll fail
to build.  This is because they set S to WORKDIR and when cros-workon tries
to do a git checkout, it fails because WORKDIR already exists.

The initial value doesn't even really matter because cros-workon will make
sure that the dir is initialized correctly, or that S is reset to a valid
location.

Thus add an explicit sanity check to prevent the uncommon case from being
broken and no one noticing.

BUG=chromium-os:33327
TEST=`emerge-stumpy xorg-conf` works w/out changes in src/platform/
TEST=`emerge-stumpy xorg-conf` works w/changes in src/platform/
TEST=`cros_workon start xorg-conf && emerge-stumpy xorg-conf` works
TEST=`rm -rf src/platform/xorg-conf && emerge-stumpy xorg-conf` works
TEST=`cbuildbot {x86,amd64,arm}-generic-full` worked

Change-Id: I990fca978e9dee0b3c8d67f024c6dcfbb5be40c9
Reviewed-on: https://gerrit.chromium.org/gerrit/43552
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>