cros-devutils: do not feed --test flag to *_unittest.py files
Few of the existing files in platform/dev/host/lib have embedded test
code in them, which is triggered by a --test argument. The ebuild was
written to run all files in that directory with a --test argument.  The
problem is that dedicated *_unittest.py files, which simply run
unittest.main(), do not take a --test argument and will fail the
invocation.  This CL distinguishes between the two categories, invoking
*_unittest.py files as it is intended.

(In general, the practice of bundling unit tests with core logic in
library files is debatable, hence my reluctance to just have all tests
in that directory bundled in this way.)

BUG=None

TEST=Ebuild runs unit tests correctly for both types.

Change-Id: I4f23c961e4ed3d56bffe719d44927bd4a45ae8b3
Reviewed-on: https://gerrit.chromium.org/gerrit/43719
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>