BlueZ 4.99: Add GetCachedServices to device API.
This fix implements a new GetCachedServices() function with similar
functionality as DiscoverServices(), except it only retrieves information
from an internal cache.

BUG=chromium-os:38410
TEST=manual test (see below)

Manual testing
--------------
Replace "12345" and "C0_0F_FE_C0_0F_FE" for the results from the first two
commands. You need a paired device in order to test this.

$ dbus-send --system --dest=org.bluez --print-reply / org.bluez.Manager.DefaultAdapter
method return sender=:1.1541 -> dest=:1.1567 reply_serial=2
   object path "/org/bluez/12345/hci0"

$ dbus-send --system --dest=org.bluez --print-reply /org/bluez/12345/hci0 org.bluez.Adapter.GetProperties | grep '"Devices"' -A4
         string "Devices"
         variant             array [
               object path "/org/bluez/12345/hci0/dev_C0_0F_FE_C0_0F_FE"
            ]
      )

The output of the following two commands must be the same when the device is
in range and with the radio on (for example, is connected). The output of the
first command will not change even if the device is off, while the second
command must fail in that case. A possible difference in the order of the
output array is ok.
$ dbus-send --system --dest=org.bluez --print-reply /org/bluez/12345/hci0/dev_C0_0F_FE_C0_0F_FE org.bluez.Device.GetCachedServices string:
$ dbus-send --system --dest=org.bluez --print-reply /org/bluez/12345/hci0/dev_C0_0F_FE_C0_0F_FE org.bluez.Device.DiscoverServices string:

Change-Id: I52ae1860e7e68262477dacfc06f76635f1502eb4
Reviewed-on: https://gerrit.chromium.org/gerrit/44267
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>