Seperate the depthcharge bits out in chromeos-bootimage.
The logic in src_compile was getting too complicated, and it has always been a
little awkward to shoehorn depthcharge into the image by tweaking the code
originally intended for U-Boot only.

This change splits out the depthcharge and non-depthcharge parts of
src_compile into two new functions, src_compile_uboot and
src_compile_depthcharge, and decides which to use in the original src_compile
based on whether the depthcharge USE flag is set.

Ultimately we'll want to actually seperate the depthcharge logic out entirely
into its own ebuild which is parallel to and substitutable for the U-Boot one,
but that's a larger change with farther reaching implications that can wait
for now. It would also be nice to split out the memtest based image similarly,
and to make it easier to get at the default locations of things like the
bitmaps without having to redefine them everywhere.

BUG=chrome-os-partner:8339
TEST=Built depthcharge, U-Boot, and memtest images for Link and booted each.
Verified that depthcharge and U-Boot started ChromeOS successfully, and that
memtest started, although it crashed for some reason. The crash also happened
without this change.
BRANCH=None

CQ-DEPEND=CL:*32999

Change-Id: I234470c4fbb6a96fdfe2be8c601d826f04df52c0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/44163
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>