Move mount-encrypted utility from vboot_reference to cryptohome
The mount-encrypted utility sets up the encrypted directories for Chrome OS
at boot time. It uses some of the TPM-related libraries from
vboot_reference, but it isn't really part of the verified boot process
itself so I'm moving it into the cryptohome repo where it belongs.

BUG=chromium-os:39264
BRANCH=none
TEST=auto
CQ-DEPEND=CL:44301, CL:44303

This is just refactoring. After all the CLs are in, the
platform_EncryptedStateful autotest should continue to pass as before.

Change-Id: I7e14afe9b34819e1ab8dbe195c9de87c9ec4d6f1
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44302
Reviewed-by: Kees Cook <keescook@chromium.org>