Install futility correctly on recovery image
The futility wrapper is (currently) a single symlinked executable that execs
the real executable from an old_bins subdirectory. When we create the
initramfs for the recovery kernel, we need to maintain that directory
structure. This change does that.

Note that this only works if we're actually building and installing the
futility wrapper in the rootfs in the first place, so there's a matching CL
for vboot_reference that has to enable that.

Also note that eventually futilty will no longer require the subdirectory.
When that happens, we can undo this change.

BUG=chromium-os:37062
BRANCH=none
TEST=manual

./build_packages --board link
./build_image --board link
./mod_image_for_recovery.sh --board link

Install recovery image onto USB, try it.

Change-Id: I802207d9320549513db3a3c2e00e1a13ec595da4
Reviewed-on: https://gerrit.chromium.org/gerrit/44864
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>