bluez - create unstable ebuild for 5.3
BlueZ 5.3 provides support for Bluetooth 4.0 BR+LE (Bluetooth Smart),
but requires a newer kernel (or the later stack backported, CL in
review) and changes the D-Bus APIs so is not compatible with our
existing userland.

This can be installed using the following commands in the chroot:

  emerge-${BOARD} --unmerge bluez bluez-hcidump bluez-test
  ACCEPT_KEYWORDS="~amd64 ~arm ~x86" USE=readline emerge-${BOARD} bluez

And then on your device:

  emerge --unmerge bluez bluez-hcidump bluez-test
  gmerge -n bluez
  reboot

After restarting, the session manager will automatically pass the
--enable-experimental-bluetooth flag to Chromium to switch the
Bluetooth UI implementation to the new APIs (currently only a stub
implementation).

BUG=chromium-os:38832
TEST=see above

Change-Id: Id62f2dad078039a87f72783096b6f2436d19ad57
Reviewed-on: https://gerrit.chromium.org/gerrit/45084
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>