Fix Telemetry on Chromeos Packaging.
The bootstrap_deps file was moved in the chrome repo, thus breaking
how we determine which files to copy over and package.

Now we only depend on the run_multipage_benchmark script to get the
telemetry_deps.

This change also hardcodes in the path for perf/data until telemetry
switches over to using Isolate.

BUG=chromium-os:39827
TEST=trybot build.

Change-Id: Id27ef05ad0bfb76d0431f0472b06c8fd8d6859be
Reviewed-on: https://gerrit.chromium.org/gerrit/45161
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>