Use deploy_chrome in the Chrome ebuild.
Now we have one source of truth and one code path for deploying Chrome
artifacts.  Also prevents changes to the ebuild from inadvertently
breaking deploy_chrome.

Introduces a dependency between a version of Chrome and the version of
chromite needed to deploy it.  This can become a problem when an
artifact is renamed, etc., and a change is needed to deploy_chrome in
chromite, but the change needs to land right when Chrome is uprevved but
not any earlier.  I'll be solving this problem shortly by adding a
pinned chromite to the Chrome DEPS file, and using the pinned
deploy_chrome in the ebuild.

BUG=None
TEST=- Trybots: x86-generic-full, lumpy-release (HWTest),
                amd64-generic-full, mario-release (HWTest),
                daisy-release (tested local deploy as well),
                x86-generic-chromium-pfq,
                lumpy-chrome-pfq, amd64-generic-chromium-pfq,
                daisy-chromium-pfq.
     - Verified that the files deployed and their permissions are
       identical before and after using deploy_chrome.
CQ-DEPEND=CL:45818

Change-Id: I61d4d43266c5ecdd14568104ebcfdabd774ec330
Reviewed-on: https://gerrit.chromium.org/gerrit/45516
Commit-Queue: Ryan Cui <rcui@chromium.org>
Reviewed-by: Ryan Cui <rcui@chromium.org>
Tested-by: Ryan Cui <rcui@chromium.org>