Make the cros-firmware.eclass compatible with depthcharge's naming convention.
Depthcharge produces several images, current image.bin, image.rw.bin, and
image.net.bin, but none have the form image-*.bin which is what the
cros-firmware eclass expects when it builds its local updaters. This change
makes it so that when the depthcharge use flag is set, it looks for a single
file, image.bin, to create a single updater, updater.sh.

It's quite possible that components farther down the updater pipeline will
need to be adjusted to expect an updater.sh instead of an updater-*.sh, but I
wasn't able to figure out what the next step is.

BUG=chromium:174631
TEST=Built chromeos-bootimage and then chromeos-firmware-daisy with depthcharge
turned on for both and then off for both. Saw that the emerge finished
successfully in both cases, and that the expected updaters were installed.
BRANCH=None

Change-Id: Ib463b62ce54a02fb4c780923747f61c640990557
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/45683
Reviewed-by: Peter Mayo <petermayo@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>