In-memory LLVM compiler cache.
It can take hundreds of milliseconds per shader to generate code with
LLVM.  Mesa's existing caching is per shader per context which does
not help chrome when it uses the same shaders in different contexts.
This patch set places a cache before the LLVM compiler which works for
all contexts.

Patch 4 just moves the stuff to be cached to its own struct; there is
no functional change.  Patch 5 adds the cache.
Cache items are small (on the order of 10K) so it makes sense not
to delete them as soon as they are no longer used.  Patch 6 adds
keeping some unused items in case they are needed later, like when
a tab is re-opened.

BUG=chromium:130005
TEST=saves a few seconds re-opening webgl aquarium

Change-Id: Ibf1ff500ff58153bb39262851b2e001f8ce0e2a5
Reviewed-on: https://gerrit.chromium.org/gerrit/45705
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>
Tested-by: Frank Henigman <fjhenigman@chromium.org>