BlueZ 5: Autopair plugin with the same 4.99 functionality
Implements an autopair plugin for providing common pincodes to dumb devices
(mice, keyboards, etc). It also generates a random pincode for keyboards
and sends it back to the agent.

To implement this plugin, a bonding retry mechanism is introduced, totally
transparent to the user, allowing the plugin to try one or more pin codes
and then ask the agent for another pincode.

BUG=chromium:196734
TEST=manual test, see below.

Manual test

1. Install bluez-5.3-r4 unstable package (see instructions on
https://gerrit.chromium.org/gerrit/45084 ).
2. Run "bt_console" on crosh and run the following commands (lines prefixed
with [bluetooth]#  are the comands to type):

[bluetooth]# power on
Changing power on succeeded
[bluetooth]# agent on
Agent registered
[bluetooth]# scan on
Discovery started

3. Put a mouse in discovery mode, and wait until it appears to start the
pairing. Here 00:11:22:AA:BB:CC is the device mac address.

[NEW] Device 00:11:22:AA:BB:CC Some Mouse
[bluetooth]# pair 00:11:22:AA:BB:CC
Attempting to pair with 00:11:22:AA:BB:CC
[CHG] Device 00:11:22:AA:BB:CC Paired: yes
Pairing successful

The pairing should work without asking for a pin code.

4. Put a keyboard in discovery mode and try to pair in the same way.

[NEW] Device 00:11:22:AA:BB:CC Some Keyboard
[bluetooth]# pair 00:11:22:AA:BB:CC
Attempting to pair with 00:11:22:AA:BB:CC
[agent] PIN code: 839775
[CHG] Device 00:11:22:AA:BB:CC Connected: yes
[CHG] Device 00:11:22:AA:BB:CC Paired: yes
[CHG] Device 00:11:22:AA:BB:CC Modalias: usb:v045Ep8502d011C
Pairing successful

This time the agent (in red) asks you to type the given pin code in the
bluetooth keyboard.

Change-Id: Ibfc7dc4bcdd55a8de21dfadf4f9dca620b9747b2
Reviewed-on: https://gerrit.chromium.org/gerrit/46319
Commit-Queue: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>