llvm: redo masking
The current keyword/mask settings allow for the release llvm to be
installed into the target (with the snapshot one masked out).  But
it also allows the release llvm to be installed into the sdk which
is not what we want.

It mostly works atm because we also install clang into the sdk and
it depends on a snapshot version of llvm.

Tweak how we control which version of llvm gets installed:
 - snapshot llvm gets stable keywords
 - release llvm gets unstable keywords
 - we package.mask snapshot llvm for CrOS boards
 - we package.keyword llvm for CrOS boards

This way the sdk only sees the snapshot version via stable keywords
while the CrOS boards only see the release version.

BUG=chromium:206038
TEST=`emerge -pv llvm` shows llvm-3.2_pre170392
TEST=`emerge-stump -pv llvm` shows llvm-3.2-r1

Change-Id: Ibc80735ccc75f35bd9e7ba667acd981731c9d419
Reviewed-on: https://gerrit.chromium.org/gerrit/47533
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>