x11-drivers/xf86-video-intel: BACKPORT: DRI2GetMSC: Do not send a bogus ust for when the drawable is not displayed
According to the opengl glx_sync_control spec, the Unadjusted System Time
(or UST) is a 64-bit monotonically increasing counter that is available
throughout the system:
http://www.opengl.org/registry/specs/OML/glx_sync_control.txt

Therefore, sending 0, even in this corner case, is out of spec. However,
we cannot just return FALSE here as that triggers a BadDrawable error to
be sent, and as is often the case mishandled, to the client. This results
in a certain compositor terminating, for example.

As an alternative we can use the monotonic system timestamp which in
theory should also be monotonic with the previous and subsequent vblank
times.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:223001
TEST=manually on link:
 1) attach external monitor (DP)
 2) start youtube video on external monitor
 3) uplug external monitor

Change-Id: Ie20fec8d6f6ac80f7826592445b24d36d6a5c199
Reviewed-on: https://gerrit.chromium.org/gerrit/46850
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>