cros-kernel2: Add support for wireless34 USE flag
The new wireless34 use flag is used to specify whether a 3.8 kernel
should build with the old 3.4 wifi driver/stack or with the newer 3.8 one.

It is expected that overlays that want to stay back with 3.4 drivers
set the USE flag in their make.conf, and that others will continue to
move forward.

BUG=chromium:230704
Change-Id: I43657bde746a8616c9384b695aef5b231fe4915e
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47959
Reviewed-by: Mike Frysinger <vapier@chromium.org>