coreboot: Don't hang on new config options
Until now the coreboot ebuilds would hang if a new config
option is added, instead of assuming the default value for
that option.
This CL changes that behavior to use the default value instead.
This also allows to weed out the configuration files for all boards,
keeping only the options around that we are changing.

Signed-off-by: Stefan Reinauer <reinauer@google.com>
BUG=chrome-os-partner:18638
BRANCH=none
TEST=emerge-link chromeos-coreboot still works and no functional
     changes are noticed.

Change-Id: I080a7869c2442f93c4a215a5dc3b39b131e8819f
Reviewed-on: https://gerrit.chromium.org/gerrit/48050
Commit-Queue: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Randall Spangler <rspangler@chromium.org>