BlueZ 5: Autopair plugin support for keyboards with fixed pincode.
Some keyboards like the Motorola SJYN0700A are supposed to be
connected to a phone, and thus only accept a fixed pincode
of "0000". It is possible to identify this kind of keyboards
because they reject the random generated pincode in a fraction
of a second while other keyboards wait for the user to press
ENTER on the Bluetooth keyboard or fail after a bigger timeout
in the order of few seconds.

This patch extends the autopair plugin functionality to detect
those keyboards and provide a "0000" pincode. It also includes
few minor changes to the autopair plugin reported on the BlueZ
mailing list.

This patch also includes a minor fix in the autopair plugin
seeding the PRNG from /dev/urandom at startup instead of time(0)
in each attempt.

BUG=chromium:225876
TEST=Manual test

1. Install BlueZ 5.4-r1 unstable package (see instructions on
https://gerrit.chromium.org/gerrit/45084 ).

2. On the "bt_console" crosh command, run:
power on
scan on
agent on

3. Put your Motorola SJYN0700A on discovery mode turning it
on and wait until it appears on your screen:
[NEW] Device 00:0F:F6:AA:BB:CC Motorola Bluetooth Wireless Keyboard

4. Attempt a pairing with it and it should ask you to enter a pin
code but later proceed with the pairing successfully.

pair 00:0F:F6:AA:BB:CC
Attempting to pair with 00:0F:F6:AA:BB:CC
[agent] PIN code: 040803
[CHG] Device 00:0F:F6:AA:BB:CC Connected: yes
[CHG] Device 00:0F:F6:AA:BB:CC Connected: no
[CHG] Device 00:0F:F6:AA:BB:CC Connected: yes
[CHG] Device 00:0F:F6:AA:BB:CC Paired: yes
Pairing successful

Change-Id: If042bfa0dbf473b6b86ed171adff2b4a193897c2
Reviewed-on: https://gerrit.chromium.org/gerrit/47924
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>