Enable sqlite USE flag for python and extensions USE flag of sqlite
Minijack, a factory log analysis tool, uses sqlite3 as the database.
We need to turn on the sqlite3 library of python to support it.
Enabling sqlite of python requries sqlite package and also need to
enable the extensions of sqlite.

We only want this change on chroot, but not the Chrome OS image.

BUG=chrome-os-partner:18691
TEST=Run "./build_packages --board link" passed.
TEST=Run "./build_packages --board daisy" passed.

Change-Id: I27ffaa041c07fbd2488fa9fc184e4260fcb5f08c
Reviewed-on: https://gerrit.chromium.org/gerrit/48094
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>