hostapd: Add optional configuration for urandom
Provide an ebuild option for allowing the use of /dev/urandom
instead of /dev/random for the random number generator source.
This source performs more quickly at the expense of security but
can be useful in some limited scenarios like a testbed for WPA
authentication where the the client should not be denied access
just because the random number pool is too small.

CQ-DEPEND=CL:48906
BUG=chromium:218911
TEST=USE=weak_urandom_low_security emerge-stumpy hostapd > /dev/null && strings -a /build/stumpy/usr/sbin/hostapd | grep dev.\*random

Change-Id: If2c24a205830e97b60bfe5ecb2edc45a2584fb4d
Reviewed-on: https://gerrit.chromium.org/gerrit/49020
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>