cros-kernel2.eclass: only build uImage on !device_tree overlays
On multiplatform kernels you need to pass in LOADADDR to the build
if you want to produce an uImage.  Since we're discarding the uImage
portion and only use the zImage, let's just skip that step alltogether
and target zImage instead.

Keep the old behavior around for !device_tree. I think it's probably
time to switch the default and sunset support for non-DT, but that's a
separate ball of wax.

BUG=none
TEST=check out for-next from arm-soc.git, make sure a kernel can still be built.

Change-Id: I5e2792bf4322e1e691f59bad6d182e4bdb3b7d47
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49021
Reviewed-by: Doug Anderson <dianders@chromium.org>