Add wifi testbed option to kernel eclass
Add a config fragment to the kernel eclass that enables features
for a wifi testbed appliance.  This should not be used except in
a specific controlled configuration, so add a warning when a
kernel is built with this USE flag.

BUG=chromium:218911
TEST=Compile with and without this flag.  Ensure that with the
flag enabled, a warning is emitted, and the kernel builds with
this config flag enabled
(/build/xxx/var/cache/portage/sys-kernel/chromeos-kernel/.config.)

Change-Id: I2aa8740c0d69710d1e7da8320ea0e286487305ba
Reviewed-on: https://gerrit.chromium.org/gerrit/49177
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>