Update the chromeos-bootimage dependencies for Coreboot on ARM.
When building a depthcharge based firmware image, Coreboot and the Coreboot
utilities are required even if we're building an ARM image. This was
purposefully omitted in the past because Coreboot for ARM couldn't be built
using our ebuilds and had to be copied in from an external build. With the
dependency missing, the image would build correctly even though portage hadn't
provided all the files that were being used.

BUG=chrome-os-partner:18734
TEST=Built a depthcharge and U-Boot based image for Link and Snow. Uninstalled
depthcharge and Coreboot on Snow and verified that emerging chromeos-bootimage
with the depthcharge USE flag set pulled them back in again.
BRANCH=None

Change-Id: I986b2f9496f628ee5360734faa0dee91370f5f0e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/49171
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>