Fold autotest-telemetry ebuild into the autotest-chrome ebuild.
This change removes the autotest-telemetry ebuild, and folds it
into the autotest-chrome ebuild instead.  This will greatly
simplify the conversion of autotests that depend on PyAuto to
depend on Telemetry instead, because we avoid having to deal
with file collisions that may happen when moving tests between
ebuilds.

Moreover, there does not seem to be a compelling reason to have
a separate autotest-telemetry ebuild.

This CL must land with a corresponding CL that adds a new
experimental BVT test called login_CryptohomeMountedTelemetry
that is similar to login_CryptohomeMounted, but it uses
Telemetry-based login instead of PyAuto-based login.  It will
eventually replace the PyAuto-based version once it's deemed
to be reliable enough.

CQ-DEPEND=CL:59601
BUG=chromium:236453
TEST=cbuildbot --remote --hwtest --chrome_rev=tot -g '59599 59601' lumpy-release

Change-Id: Ib11a0049ff588a148a23dda473844c0db76bd7d2
Reviewed-on: https://gerrit.chromium.org/gerrit/59599
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>