Fold autotest-telemetry ebuild into the autotest-chrome ebuild.
This change removes the autotest-telemetry ebuild, and folds it
into the autotest-chrome ebuild instead.  This will greatly
simplify the conversion of autotests that depend on PyAuto to
depend on Telemetry instead, because we avoid having to deal
with file collisions that may happen when moving tests between
ebuilds.

Moreover, there does not seem to be a compelling reason to have
a separate autotest-telemetry ebuild.

This CL must land with a corresponding CL that converts
login_CryptohomeMounted to use Telemetry-based login instead of
PyAuto-based login.

CQ-DEPEND=CL:50060
BUG=chromium:236453
TEST=cbuildbot --remote --hwtest --chrome_rev=tot -g '50060 50061' lumpy-release

Change-Id: If3a9a3b8cedede4f48adec7ddf51b6bc1f6d20c1
Reviewed-on: https://gerrit.chromium.org/gerrit/50061
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>