BlueZ: Request the device long name when only the short name is known.
update_found_devices() could attempt to update a device with new information
containing only a short verstion of the name. If the struct eir_data passed
has a short or NULL name, it will never update a previous name and will
also trigger a "HCI Command: Remote Name Request" if the previous name was
unknown.

BUG=chromium:237344
TEST=Manual test.

Manual test procedure:
1. Scan for new devices in a fresh install.
Note: if the device was already connected or paired with the
chromebook, run from a root terminal:
  # stop bluetoothd
  # rm -r /var/lib/bluetooth/*
  # start bluetoothd

2. Put a Monster ClarityHD or a Creative WP-350 Headset in discovery
mode.

3. The device MUST show up in the list with the full name in no more
than about 20s. The device COULD show up in the list with a short name
(like "Monster C" or "Creative ") for a short period of time, but it
MUST appear with the full name at some point and keep that long name.

Change-Id: If4f23a874c65f4db1e46db0f179c0ed25581d350
Reviewed-on: https://gerrit.chromium.org/gerrit/50251
Commit-Queue: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>