BlueZ: Do not retry a bonding when no pincode was requested.
Currently, a new bonding attempt will be initiated if a previous one
returns with an authentication error. This could happen when a Bluetooth
2.1 device is in discovery mode but not in pairable mode, which will
cause a bonding attempt to fail with an authentication error because
of the lack of an existing link key. The bonding retry mechanism
assumes that the authentication error was caused by the wrong pin
code provided and thus will retry with the next one.

This patch avoids a bonding attempt retry when no pin code was requested
during the bonding attempt and passes the authentication failure to
the client.

BUG=chromium:238415
TEST=Manual test.

Manual test procedure:
1. Power on a Monster ClarityHD bluetooth speaker, but DON'T put it in discovery mode. LED should be steady blue.
2. Attempt to add a new bluetooth device, the Monster ClarityHD will show up in the list (even if it's not in pairing mode!)
3. Attempt to pair with it.
4. The connecting message could stay for a short amount of time, but then an error message should be seen.

Change-Id: I33af9bbc1131e042e99017e81eb68f263032a573
Reviewed-on: https://gerrit.chromium.org/gerrit/50330
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>