shill ebuild: remove unnecessary CXXFLAG
Clang's c++98 implementation requires that the copy constructor be
accessible when passing a temporary by reference. This caused problems,
because the Error class has a private copy ctor. To override this
requirement, we used -Wno-bind-to-temporary-copy.

This requirement was removed in c++11. Since we're now compiling shill
with -std=g++11 (for both g++ and clang++), we can remove this CXXFLAG.

Note that both shill and its unittest compile without trouble.
The unittest binary aborts (see chromium:238790), but that happens even
without this change.

Info on c++98 vs c++11:
http://stackoverflow.com/a/13898867/2197141
http://stackoverflow.com/a/1827049/2197141

BUG=chromium:215496
TEST=FEATURES=test USE=clang emerge-link shill

Change-Id: I86f2ac125525afbcd0b55d74397ad7c3f4248760
Reviewed-on: https://gerrit.chromium.org/gerrit/50357
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>