Use the deploy_chrome from the Chrome checkout.
Now that chromite is part of the Chrome checkout, run the deploy_chrome
from that checkout instead of from the current Chrome OS checkout.  The
benefit of doing this is if a new buildspec of Chrome requires a
non-backwards compatible change to deploy_chrome, we can commit the
fix to deploy_chrome without breaking existing Chrome OS release builds,
and then roll the DEPS for chromite in the Chrome checkout.

Another benefit is each version of Chrome will have the right
corresponding version of deploy_chrome.

BUG=None
TEST=Local, trybots

Change-Id: Iae19e36e1ac85024368fa6fc8d8893482df17910
Reviewed-on: https://gerrit.chromium.org/gerrit/50378
Reviewed-by: Ryan Cui <rcui@chromium.org>
Tested-by: Ryan Cui <rcui@chromium.org>