Disable libatmoic for gcc 4.8.
GCC 4.7 does not include an atomic library implementation as the
atomic API has not been firmly established. GCC 4.8 now includes a
libatomic implementation with the toolchain.

However this new libatmoic does not build for this time being, safely
disable it, since there is no packages in chromiumos depends on this
library.

This patch is needed for buildbots that uses gcc 4.8 as the toolchain.

TEST=cbuildbot chromiumos-sdk, lumpy-relase, arm-release
BUG=None

Change-Id: Id4de3577fbcc448c9a2926e1cd48e3c59ebaa9d2
Reviewed-on: https://gerrit.chromium.org/gerrit/50473
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Han Shen <shenhan@google.com>
Tested-by: Han Shen <shenhan@google.com>