Clean up comments in Chrome ebuild a bit.
This CL just cleans up comments. It also ensures that Chrome will be
revved. We want Chrome to be revved now so that we can find out sooner
whether there are any issues CL:50485. (I.e., so we don't find out in
the middle of the night.)

BUG=chromium:222377
CQ-DEPEND=CL:50485
TEST=Run --buildbot --debug run of chrome builders of this CL with CL:50485.

Change-Id: Ie40d1f9288655dee7f36d4b22dbac83f41a8d149
Reviewed-on: https://gerrit.chromium.org/gerrit/50701
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>