avahi: default to IPv4 for now
Defaulting to IPv6 exposed a bug in update_engine whereby URLs with
literal IPv6 addresses does not work as expected. This is essentially
a libcurl bug, see

 http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=476470
 http://curl.haxx.se/docs/knownbugs.html (item 30)

for details. It is not expected that this bug is hit in production
since Omaha usually returns URL with names and never literal IPv6
addresses. I've raised this issue internally and will work on it once
there's agreement on how to proceed. Until then, let's default to IPv4
since it's bound to be more widely tested and also consistent with how
Ubuntu, Fedora and other Linux OSes configure Avahi.

BUG=None
TEST=Checked Avahi only responds to mDNS via IPv4 and only returns A records

Change-Id: I7ba7b5e15a42294f5c45b40c63c25200f21acfed
Reviewed-on: https://gerrit.chromium.org/gerrit/50709
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>